chore(graph-layers): remove test-utils module, reorganize code
