+++ modules/react-graph-layers/src/use-graph-engine.tsx
@@ -1,5 +1,5 @@
 import {useLayoutEffect, useRef, useState} from 'react';
-import {BaseLayout, Graph, GraphEngine} from 'deck-graph-layers';
+import {BaseLayout, Graph, GraphEngine} from '@deck.gl-community/graph-layers';
 
 export const useGraphEngine = (graph: Graph, layout: BaseLayout): GraphEngine => {
   const [engine, setEngine] = useState(new GraphEngine(graph, layout));
