--- modules/test-utils/README.md
@@ -1,5 +1,0 @@
-# @deck.gl-community/test-utils
-
-vis.gl utilities for tests and examples.
-
-See [vis.gl](http://vis.gl) for documentation.
