--- modules/test-utils/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@deck.gl-community/test-utils",
-  "private": true,
-  "description": "Internal utilities for tests and examples",
-  "license": "MIT",
-  "version": "0.0.0",
-  "type": "module",
-  "sideEffects": false,
-  "types": "./dist/index.d.ts",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "require": "./dist/index.cjs",
-      "import": "./dist/index.js"
-    }
-  },
-  "scripts": {
-    "test": "vitest run",
-    "test-watch": "vitest"
-  },
-  "devDependencies": {
-    "ngraph.generators": "^20.1.0"
-  }
-}
