--- modules/test-utils/src/index.ts
@@ -1,1 +1,0 @@
-export * from './data/graphgl/index';
