+++ package.json
@@ -12,6 +12,7 @@
   },
   "type": "module",
   "workspaces": [
+    "examples/graph-layers/react-graph-gl",
     "examples/editable-layers/editor",
     "modules/*"
   ],
