+++ tsconfig.build.json
@@ -14,8 +14,7 @@
     {"path": "modules/layers"},
     {"path": "modules/react"},
     {"path": "modules/react-graph-layers"},
-    {"path": "modules/template"},
-    {"path": "modules/test-utils"}
+    {"path": "modules/template"}
   ],
   "include": ["modules/*/src"],
   "exclude": ["modules/*/test"],
