+++ tsconfig.json
@@ -22,7 +22,6 @@
       "@deck.gl-community/layers": ["modules/layers/src"],
       "@deck.gl-community/react": ["modules/react/src"],
       "@deck.gl-community/template": ["modules/template/src"],
-      "@deck.gl-community/test-utils": ["modules/test-utils/src"],
       "deck-graph-layers": ["modules/graph-layers/src"],
       "react-deck-graph-layers": ["modules/react-graph-layers/src"]
     },
